refactor : test 클래스 분리
